fix: 서브 도메인 생각해서 route에서 admin 제거
